requirements.txt
asgiref==3.4.1
certifi==2022.9.24
charset-normalizer==2.0.12
Django==3.2.15
djangorestframework==3.14.0
djangorestframework-simplejwt==4.4.0
idna==3.4
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2022.2.1
requests==2.27.1
sqlparse==0.4.3
typing-extensions==4.1.1
urllib3==1.26.12
